+++ src/gatling/simulations/config/Environment.scala
@@ -5,7 +5,7 @@ import scala.util.Properties
 
 trait Environment {
 
-  val baseUrl: String = Properties.envOrElse("HOST_NAME", "https://www.xebia.com")
+  val baseUrl: String = Properties.envOrElse("HOST_NAME", "https:/google.com")
   val loadFactor: Double = Properties.envOrElse("LOAD_FACTOR", "1").toDouble
   val durationInMinutes: Int = Properties.envOrElse("DURATION_MINUTES", "1").toInt