+++ CHANGELOG.md
@@ -4,10 +4,30 @@
 
 
 <a name="v0.1.0"></a>
-## v0.1.0 - 2020-05-24
+## [v0.1.0] - 2020-05-24
 
-- initial version
+- get and post examples
+
+
+<a name="v0.0.3"></a>
+## [v0.0.3] - 2020-05-24
+
+- not provided simulation
+
+
+<a name="v0.0.2"></a>
+## [v0.0.2] - 2020-05-24
+
+- config and feeders
+
+
+<a name="v0.0.1"></a>
+## v0.0.1 - 2020-05-24
+
 - Initial commit
 
 
-[Unreleased]: https://github.com/ik-performance/performance-blueprint-project.git/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/ik-performance/gatling-blueprint.git/compare/v0.1.0...HEAD
+[v0.1.0]: https://github.com/ik-performance/gatling-blueprint.git/compare/v0.0.3...v0.1.0
+[v0.0.3]: https://github.com/ik-performance/gatling-blueprint.git/compare/v0.0.2...v0.0.3
+[v0.0.2]: https://github.com/ik-performance/gatling-blueprint.git/compare/v0.0.1...v0.0.2